requirements.txt
python-valve
discord
mcstatus